test: 딜러 및 플레이어 테스트 추가
