+++ src/test/kotlin/blackjack/domain/model/DealerTest.kt
@@ -9,12 +9,12 @@ class DealerTest {
 
     @BeforeEach
     fun setUp() {
-        dealer = Dealer("동전", listOf(Card(Suit.HEART, Rank.ACE)))
+        dealer = Dealer(cards = listOf(Card(Suit.HEART, Rank.ACE)))
     }
 
     @Test
     fun `플레이어별 승패 여부를 반환한다`() {
-        val player1 = Player("A", listOf(Card(Suit.HEART, Rank.TWO))) //  2점
+        val player1 = Player("A", listOf(Card(Suit.HEART, Rank.TWO))) // 2점
         val player2 = Player("B", listOf(Card(Suit.HEART, Rank.ACE))) // 11점
         val player3 = Player("C", listOf(Card(Suit.HEART, Rank.ACE), Card(Suit.HEART, Rank.KING))) // 21점
         val verdicts: Map<Player, Result> = dealer.getPlayerResult(listOf(player1, player2, player3))
@@ -29,7 +29,7 @@ class DealerTest {
 
     @Test
     fun `딜러의 승패 횟수를 반환한다`() {
-        val player1 = Player("A", listOf(Card(Suit.HEART, Rank.TWO))) //  2점
+        val player1 = Player("A", listOf(Card(Suit.HEART, Rank.TWO))) // 2점
         val player2 = Player("B", listOf(Card(Suit.HEART, Rank.ACE))) // 11점
         val player3 = Player("C", listOf(Card(Suit.HEART, Rank.ACE))) // 11점
         val player4 = Player("D", listOf(Card(Suit.HEART, Rank.ACE))) // 11점
@@ -40,5 +40,17 @@ class DealerTest {
         val verdicts: Map<Result, Int> = dealer.getDealerResults(playerVerdicts)
         val actual: Map<Result, Int> = mapOf(Result.WIN to 1, Result.LOSE to 2, Result.DRAW to 3)
         assertThat(verdicts).isEqualTo(actual)
+    }
+
+    @Test
+    fun `딜러의 점수가 16 이하면 히트할 수 있다`() {
+        dealer = Dealer(cards = listOf(Card(Suit.HEART, Rank.SIX), Card(Suit.HEART, Rank.KING)))
+        assertThat(dealer.canHit()).isTrue()
+    }
+
+    @Test
+    fun `딜러의 점수가 16 초과면 히트할 수 없다`() {
+        dealer = Dealer(cards = listOf(Card(Suit.HEART, Rank.QUEEN), Card(Suit.HEART, Rank.KING)))
+        assertThat(dealer.canHit()).isFalse()
     }
 }
